fw: Use range pattern for counter field max.
This shrinks the release firmware size by quite a bit as it removes a path to
panic. It seems now the compiler can tell the range checks in set_counter() will
always be met.

(As the getters in dbc_codegen structs don't have any range bounds, there's no
guarantee the old counter value wasn't out of range, which is odd as usually
the field mask actually limits it as well...)